Put apolloClient in App component state.
I had an issue with my application not using the cache at all. Turned out it was because a new Apollo client (and a new empty cache) is created on every render if you do it as written in this tutorial. Putting the client in component state with `useState` solved the problem.